Move the adversarial review example under ccflow.examples
The example belongs beside the other worked examples rather than at the top of
the package, where it read as a first-class subsystem.

ccflow.ai becomes ccflow.examples.ai and its tests move to
ccflow/tests/examples/ai, matching where the calculator, ETL and TPC-H examples
and their tests already live. Dotted paths in the configs, the module entry
point and the documentation follow. Packaging is unchanged: the wheel declares
the whole ccflow package, so the config tree and persona files ship as before.

Also isolates the registry in the example's tests. They load into the global
root registry, so each now starts and finishes with a cleared one; without that
they pass alone and interfere in a suite, in both directions.
