+++ ccflow/examples/ai/config/base.yaml
@@ -5,18 +5,18 @@
 #
 # Run the whole graph offline (TestModel, no credentials, checks off because generated placeholder
 # text cannot satisfy a real-quotation check):
-#   python -m ccflow.ai checks=off
+#   python -m ccflow.examples.ai checks=off
 # Run it against a real model, with the mechanical constraints enforced:
-#   python -m ccflow.ai model=openai
+#   python -m ccflow.examples.ai model=openai
 # Swap one axis at a time:
-#   python -m ccflow.ai spec=haiku rubrics=strict profile=steelman
+#   python -m ccflow.examples.ai spec=haiku rubrics=strict profile=steelman
 # Judge both presentation orders and report only what survives both:
-#   python -m ccflow.ai task=counterbalanced
+#   python -m ccflow.examples.ai task=counterbalanced
 # Write each stage to disk and chain the single-node tasks:
-#   python -m ccflow.ai task=produce publisher=json
-#   python -m ccflow.ai task=advocate publisher=json position=pro
-#   python -m ccflow.ai task=advocate publisher=json position=con
-#   python -m ccflow.ai task=arbitrate publisher=json
+#   python -m ccflow.examples.ai task=produce publisher=json
+#   python -m ccflow.examples.ai task=advocate publisher=json position=pro
+#   python -m ccflow.examples.ai task=advocate publisher=json position=con
+#   python -m ccflow.examples.ai task=arbitrate publisher=json
 
 defaults:
   - _self_
