+++ ccflow/examples/ai/config/lessons/house_style.yaml
