+++ ccflow/examples/ai/config/lessons/producer_only.yaml
