+++ ccflow/examples/ai/config/rubrics/poem.yaml
