+++ ccflow/examples/ai/config/rubrics/strict.yaml
