+++ ccflow/examples/ai/config/spec/haiku.yaml
