+++ ccflow/examples/ai/config/task/rebuttal_counterbalanced.yaml
@@ -3,7 +3,7 @@
 # Both defences at once: a rebuttal round, then arbitration in both presentation orders. The
 # strongest configuration here, and the slowest -- seven model calls.
 #
-#   python -m ccflow.ai task=rebuttal_counterbalanced
+#   python -m ccflow.examples.ai task=rebuttal_counterbalanced
 #
 # A `decisive: true` verdict here means the winner survived reading the opposing case *and* survived
 # being shown in either slot. That is a materially different claim from a single-order first-round
@@ -12,13 +12,13 @@
 task_name: verdict
 
 produce:
-  _target_: ccflow.ai.ProduceModel
+  _target_: ccflow.examples.ai.ProduceModel
   enforce_checks: ${enforce_checks}
   spec: ${poem_spec}
   session: ${producer_session}
 
 pro_opening:
-  _target_: ccflow.ai.AdvocateModel
+  _target_: ccflow.examples.ai.AdvocateModel
   enforce_checks: ${enforce_checks}
   position: pro
   produce: ${produce}
@@ -26,7 +26,7 @@ pro_opening:
   session: ${pro_session}
 
 con_opening:
-  _target_: ccflow.ai.AdvocateModel
+  _target_: ccflow.examples.ai.AdvocateModel
   enforce_checks: ${enforce_checks}
   position: con
   produce: ${produce}
@@ -34,7 +34,7 @@ con_opening:
   session: ${con_session}
 
 pro_rebuttal:
-  _target_: ccflow.ai.RebuttalModel
+  _target_: ccflow.examples.ai.RebuttalModel
   enforce_checks: ${enforce_checks}
   position: pro
   own: ${pro_opening}
@@ -44,7 +44,7 @@ pro_rebuttal:
   session: ${pro_rebuttal_session}
 
 con_rebuttal:
-  _target_: ccflow.ai.RebuttalModel
+  _target_: ccflow.examples.ai.RebuttalModel
   enforce_checks: ${enforce_checks}
   position: con
   own: ${con_opening}
@@ -54,13 +54,13 @@ con_rebuttal:
   session: ${con_rebuttal_session}
 
 task:
-  _target_: ccflow.ai.DebateModel
+  _target_: ccflow.examples.ai.DebateModel
   publisher: ${publisher}
   arbiter:
-    _target_: ccflow.ai.CounterbalancedArbitrateModel
+    _target_: ccflow.examples.ai.CounterbalancedArbitrateModel
     enforce_checks: ${enforce_checks}
     forward:
-      _target_: ccflow.ai.ArbitrateModel
+      _target_: ccflow.examples.ai.ArbitrateModel
       enforce_checks: ${enforce_checks}
       rubric: ${arbitration_rubric}
       artifact: ${produce}
@@ -69,7 +69,7 @@ task:
         - ${pro_rebuttal}
         - ${con_rebuttal}
     reverse:
-      _target_: ccflow.ai.ArbitrateModel
+      _target_: ccflow.examples.ai.ArbitrateModel
       enforce_checks: ${enforce_checks}
       rubric: ${arbitration_rubric}
       artifact: ${produce}
