+++ ccflow/examples/ai/personas/con.md
