+++ ccflow/examples/ai/personas/skeptic.md
