+++ docs/wiki/reference/Reference.md
@@ -7,7 +7,7 @@ Technical description of the pieces `ccflow` provides. These pages are for looki
 - **[Built-in Models](Built-in-Models)** — the models, publishers, and evaluators that ship with `ccflow`, plus cache-key derivation.
 - **[Flow Model](Flow-Model)** — the `@Flow.model` API for defining callable models from plain Python functions.
 - **[Reporting](Reporting)** — lifecycle events, reporter sinks, reporting evaluators and models, dry-run planning, and OpenTelemetry integration.
-- **[Adversarial Review](Adversarial-Review)** — the optional `ccflow.ai` example: agents as graph nodes, typed hand-offs, and constraints enforced in code rather than by a model.
+- **[Adversarial Review](Adversarial-Review)** — the optional `ccflow.examples.ai` example: agents as graph nodes, typed hand-offs, and constraints enforced in code rather than by a model.
 
 > [!NOTE]
 > Some items are marked *Coming Soon!* — they are part of the design and are in the process of being open-sourced.
